# api/openapi/shift.yaml
openapi: 3.1.0
info:
  title: SHIFT Control Plane API
  version: 0.1.0
  description: |
    Fleet metadata and orchestration API for SHIFT. Raw computational state,
    checkpoint chunks, and workload keys remain on mutually authenticated agents.
servers:
  - url: http://127.0.0.1:8090
    description: Local development server
...
paths:
...
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Rotated session tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/organizations/{organizationID}/storage:
    parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
    get:
      tags: [Storage]
      security: [{ bearerAuth: [] }]
      operationId: getStorageStatus
      description: |
        Platform-hosted checkpoint storage status. Usage is metered by
        reconciling actual object bytes in the organization's bucket prefix —
        never from agent claims — and refreshes synchronously when the stored
        figures are stale. When hosting is disabled the response is
        `{enabled: false}` with no location fields.
      responses:
        '200':
          description: Storage status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageStatus' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SHIFT access token
  parameters:
    OrganizationID:
      name: organizationID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Authentication failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authorization failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string, writeOnly: true }
    SessionTokens:
      type: object
      required: [session_id, access_token, refresh_token, token_type, expires_at]
      properties:
        session_id: { type: string, format: uuid }
        access_token: { type: string, writeOnly: true }
        refresh_token: { type: string, writeOnly: true }
        token_type: { type: string, enum: [Bearer] }
        expires_at: { type: string, format: date-time }
...
    StorageStatus:
      type: object
      required: [enabled, used_storage_bytes, max_storage_bytes, over_quota]
      properties:
        enabled: { type: boolean, description: False when this control plane does not host storage; location fields are then absent }
        endpoint: { type: string, description: Object-store endpoint handed to agents }
        bucket: { type: string }
        prefix: { type: string, description: The organization's exclusive key prefix, "org/{organizationID}/" }
        used_storage_bytes: { type: integer, format: int64, minimum: 0, description: Reconciled from actual bucket bytes }
        max_storage_bytes: { type: integer, format: int64, description: -1 means unlimited }
        over_quota: { type: boolean }
        last_reconciled_at: { type: string, format: date-time }
...